requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
pymongo
requests
google-generativeai
